+++ src/app/pages/settings/settings.module.ts
@@ -14,8 +14,8 @@ import {MatTreeModule} from '@angular/material/tree';
 import {CdkTreeModule} from '@angular/cdk/tree';
 import {NgxFilesizeModule} from 'ngx-filesize';
 import {IonicModule} from '@ionic/angular';
-
-export {SettingsPageComponent};
+import {SettingsMenuComponent} from './settings-menu/settings-menu.component';
+import {SettingsAppearanceImagesComponent} from './settings-appearance/settings-appearance-images/settings-appearance-images.component';
 
 @NgModule({
   declarations: [
@@ -26,6 +26,8 @@ export {SettingsPageComponent};
     SettingsAboutComponent,
     SettingsVoiceInputComponent,
     SettingsVoiceOutputComponent,
+    SettingsMenuComponent,
+    SettingsAppearanceImagesComponent,
   ],
   imports: [
     CommonModule,
@@ -38,6 +40,6 @@ export {SettingsPageComponent};
     IonicModule,
   ],
   bootstrap: [SettingsPageComponent],
-  exports: [SettingsAppearanceComponent],
+  exports: [SettingsAppearanceComponent, SettingsAppearanceImagesComponent],
 })
 export class SettingsPageModule {}