+++ src/app/pages/translate/translate.module.ts
@@ -16,21 +16,12 @@ import {MatTooltipModule} from '@angular/material/tooltip';
 import {NgxsModule} from '@ngxs/store';
 import {TranslateState} from '../../modules/translate/translate.state';
 import {MatDialogModule} from '@angular/material/dialog';
+import {NtkmeButtonModule} from '@ctrl/ngx-github-buttons';
 
 const routes = [
   {
     path: '',
     component: TranslateComponent,
-  },
-  {
-    path: 'settings',
-    outlet: 'dialog',
-    component: LazyDialogEntryComponent,
-  },
-  {
-    path: 's', // to prevent the settings from loading on page load, adding one level of route (i.e. s/offline)
-    outlet: 'settings',
-    loadChildren: () => import('../settings/settings.module').then(m => m.SettingsPageModule),
   },
 ];
 
@@ -53,6 +44,7 @@ const components = [SendFeedbackComponent, TranslateInputButtonComponent, Transl
     MatDialogModule,
     RouterModule.forChild(routes),
 
+    NtkmeButtonModule,
     ...componentModules,
   ],
   declarations: components,